Move url.py & view.py to more precise folder
