+++ src/reviews/views.py
@@ -2,10 +2,14 @@ import collections
 import json
 import random
 
+from django.conf import settings
+from django.conf.global_settings import DATETIME_INPUT_FORMATS
+from django.core.exceptions import ValidationError
 from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.db.models import Count
 from django.http import Http404
 from django.urls import reverse
+from django.shortcuts import redirect, render
 from django.views.generic import ListView, UpdateView
 
 from core.utils import SequenceQuerySet
@@ -15,6 +19,10 @@ from .context import reviews_state
 from .forms import ReviewForm
 from .models import REVIEW_REQUIRED_PERMISSIONS, Review, TalkProposalSnapshot
 
+from registry.helper import reg
+
+import pytz
+import datetime
 
 class ReviewableMixin:
     def dispatch(self, request, *args, **kwargs):
@@ -327,3 +335,67 @@ class ReviewEditView(ReviewableMixin, PermissionRequiredMixin, UpdateView):
         if query_string:
             return url + '?' + query_string
         return url
+
+def review_stages(request):
+    current_review_stages_setting = {}
+    review_stages_list = [
+        'Call for Proposals',
+        'Locked (proposal editing and reviewing disabled)',
+        'First Round Review', 'Modification Stage', 'Second Round Review',
+        'Internal Decision', 'Announcement of Acceptance'
+    ]
+    review_stages_var = [
+        'proposals.creatable', 'proposals.editable', 'proposals.withdrawable',
+        'reviews.visible.to.submitters', 'reviews.stage',
+        'proposals.disable.after'
+    ]
+
+    if request.method == 'POST':
+        date_time_obj = date_preprocess(
+            DATETIME_INPUT_FORMATS, request.POST['proposals.disable.after'])
+        tz_selectd = pytz.timezone(request.POST['review_timezone'])
+        loc_dt = tz_selectd.localize(date_time_obj).strftime(
+            '%Y-%m-%d %H:%M:%S%z')
+
+        for tag in review_stages_var:
+            key = settings.CONFERENCE_DEFAULT_SLUG + '.' + tag
+            if (tag == 'proposals.disable.after'):
+                value = loc_dt
+            elif (tag == 'reviews.stage'):
+                value = int(request.POST[tag])
+            else:
+                value = request.POST[tag]
+            reg[key] = value
+
+        messages.info(request, 'This setting has been changed successfully.')
+    else:
+        for tag in review_stages_var:
+            key = settings.CONFERENCE_DEFAULT_SLUG + '.' + tag
+            value = reg.get(key, '')
+            # Django template language does not support dictionary keys containing "."
+            if "." in tag:
+                tag = tag.replace(".", "_")
+            current_review_stages_setting[tag] = value
+            print(tag)
+            print(value)
+
+    return render(
+        request, 'reviews/review_stages.html', {
+            'timezones': pytz.common_timezones,
+            'review_stages_list': review_stages_list,
+            'current_review_stages_setting': current_review_stages_setting,
+            **reviews_state()._asdict()
+        })
+
+
+def date_preprocess(DATETIME_INPUT_FORMATS, value):
+    # Add defined datetime formatx
+    DATETIME_INPUT_FORMATS += ['%Y-%m-%dT%H:%M:%S', '%Y-%m-%dT%H:%M']
+    value = value.strip()
+    # Try to strptime against each input format.
+    for format in DATETIME_INPUT_FORMATS:
+        try:
+            return datetime.datetime.strptime(value, format)
+        except (ValueError, TypeError):
+            continue
+    raise ValidationError("Please input valid date format : " + "%Y-%m-%dT%H:%M")
